requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
python-jose[cryptography]
passlib[pbkdf2_sha256]
python-dotenv
python-multipart
